# Cleanup of trailingBracket function