Cleaned up `trailingBracket` and renamed it to the more appropriate (in my opinion) `nestedStructureEnd`.