@buger 2016-03-27
This changes cause 10% slowdown, i'm not sure why yet, maybe switch in Go slower then if, and i will try to figure out. But i think it make sense to add policy, that when you submit any change, you should provide benchmark before and after, and if after significantly more then before, there should be some real reasons to merge it.