@daboyuka 2016-03-27
That's a good policy, I'll do that in the future. It's very strange that it's slower, though...I'll take a look as well.