@daboyuka 2016-03-27
Huh, you're right. While it's something like 7% faster on the Large benchmark, it's 10% slower on the others. I don't think this is a `switch` statement thing, likely just some common code path that is slightly more complex now. I'll check it out.