@buger 2016-03-27
I also going to rollback string escape PR because it caused even more slowdown:

Befor PR
```
BenchmarkJsonParserLarge 	  100000	     74091 ns/op	      32 B/op	       2 allocs/op
BenchmarkJsonParserMedium	  500000	     13395 ns/op	      32 B/op	       2 allocs/op
BenchmarkJsonParserSmall 	 5000000	      1206 ns/op	       0 B/op	       0 allocs/op
```

Now:
```
BenchmarkJsonParserLarge 	  100000	    115964 ns/op	      32 B/op	       2 allocs/op
BenchmarkJsonParserMedium	g  500000	     18231 ns/op	      32 B/op	       2 allocs/op
BenchmarkJsonParserSmall 	 5000000	      1419 ns/op	       0 B/op	       0 allocs/op
```

And looks like all happens because of switch.