@daboyuka 2016-03-27
You ran the benchmark a few times right? In my experience it's not uncommon for the first run after a build to be significantly slower.