@buger 2016-03-27
Yep, i usually run them multiple times with bechtime 5s to reduce noises 

```
go test -test.benchmem -bench JsonParser ./benchmark/ -benchtime 5s -v
```