@daboyuka 2016-03-27
Right, I'm using the same benchtime.

I fixed the performance issue with this PR by eliminating the `default:` case. I'll push it in a moment.