@buger 2016-03-27
I also investigated `stringEnd` changes, and it could be that additional escaped variable and its checked may cause the slowdown.