@daboyuka 2016-03-27
That's likely. Unfortunately, I think some additional checking is needed for correctness; see the test case I submitted along with that PR. Maybe the fixed version can be optimized better, though (we should take any further discussion to that PR though).