@daboyuka 2016-03-27
Nevermind, there's still some overhead with the new trailingBracket/nestedStructureEnd function. Actually, it's faster for the Large case still, but slightly slower on the Medium/Small (which are more representative of my particular real-life use case).

I can update these PRs to use ifs next time I sit down to work on this.

Before PR:
```
BenchmarkJsonParserLarge	  100000	     58867 ns/op
BenchmarkJsonParserMedium	 1000000	     10192 ns/op
BenchmarkJsonParserSmall	10000000	      1051 ns/op
```

After PR:
```
BenchmarkJsonParserLarge	  200000	     52972 ns/op
BenchmarkJsonParserMedium	 1000000	     10516 ns/op
BenchmarkJsonParserSmall	10000000	      1082 ns/op
```