@daboyuka 2016-03-27
Right, I will push momentarily.

By the way, those earlier benchmarks were using Go 1.4. I just tried using Go 1.5, and wow, something is really wrong with the Go 1.5 compiler in this case...
Before:
```
BenchmarkJsonParserLarge-8 	  200000	     52000 ns/op
BenchmarkJsonParserMedium-8	  500000	     14771 ns/op
BenchmarkJsonParserSmall-8 	10000000	      1030 ns/op
```

After:
```
BenchmarkJsonParserLarge-8 	  200000	     53395 ns/op
BenchmarkJsonParserMedium-8	 1000000	      9986 ns/op
BenchmarkJsonParserSmall-8 	10000000	      1027 ns/op
```

Note that Medium got about 50% slower transitioning from Go 1.4 to 1.5...rebuilt and reran the benchmarks several times to be sure. It could be a Mac-only thing though.

Bottom line is that benchmarks are finicky things at this low level, best not to put _too_ much stock in small changes :).