@buger 2016-03-27
Worth noticing that i run all benchmarks on 1.6 as it current stable version.