@daboyuka 2016-03-27
In a few days we will have upgraded our build/deploy chain to work on Go 1.6 as well, since GAE just upgraded. So, I will use that version soon. My understanding is not a great deal has changed compiler-optimization-wise, though.