@buger 2016-03-27
Yeah, i prefer to see all the commits without rebasing.